requirements.txt
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
openai
python-dotenv
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
python-dateutil==2.8.2
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-bootstrap5==1.0.0
flask-wtf==1.1.1
flask-paginate==2022.1.8
gunicorn==21.2.0
